♻️refactor: export 방식
