+++ src/app/user/[id]/setting/UserSetting.view.tsx
@@ -4,7 +4,7 @@ import * as S from "./UserSetting.styles";
 import * as GS from "@/styles/GlobalStyle";
 import { NAV_HEIGHT } from "@/styles/theme";
 
-import Divider from "@/components/Divider/Divider";
+import { Divider } from "@/components/Divider";
 
 import { UserSettingAccount, UserSettingMember, UserSettingOptions } from "./components";
