+++ src/components/Divider/index.ts
@@ -1,0 +1,1 @@
+export { default as Divider } from "./Divider";
